backend-engine/pyproject.toml
[project]
name = "vulnerable-banana"
version = "0.1.0"
description = "Security comics generator backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pydantic-ai>=0.0.20",
    "google-genai>=1.0.0",
    "google-cloud-storage>=2.14",
    "httpx>=0.26",
    "python-multipart>=0.0.6",
    "pillow>=10.2",
    "structlog>=24.1",
    "pyhumps>=3.8",
    "tenacity>=8.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
    "mypy>=1.8",
]

[tool.hatch.build.targets.wheel]
packages = ["src/vulnerable_banana"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
